refactor profile header to be a top level component
