--- frontend/components/Profile/ProfileHeader/ProfileHeader.less
@@ -1,52 +1,0 @@
-//header component
-.Profile_profileHeader {
-    border-radius: 10px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    z-index: 500;
-    padding: 2% 5%;
-    position: absolute;
-    background: white;
-    width: 80%;
-    height: 4rem;
-    top: 2.5%;
-    left: 10%;
-    right: 10%;
-    //dropdown component
-      .profile_userDropDown {
-        width: 40%;
-        border: 1px solid black!important;
-        border-radius: 40px!important;
-        text-align: center;
-        font-size: 2rem;
-        min-width: none!important;
-        min-height: 0rem!important;
-        padding: 1rem!important;
-        display: flex;
-        justify-content: flex-end;
-        //input cursor
-          input.search {
-            margin-left: 1rem;
-          }
-          //the actual dropdown
-          .visible.menu.transition {
-            width: 100%
-          }
-          i {
-            padding: .2rem 1.2rem 2rem 16rem;
-            margin-right: 1rem;
-          }
-          //inside the drop down, each item
-          .profile_dropList {
-            i {
-              font-size: 1.3rem;
-              padding: 0rem;
-            }
-            span {
-                font-size: 1.3rem;
-            }
-          }
-      }
-    
-}
