+++ frontend/components/ProfileHeader/ProfileHeader.less
