+++ frontend/components/ProfileHeader/UserDropdown.js
@@ -9,10 +9,10 @@
 //-- Dependencies --------------------------------
 import React from 'react'
 import { Dropdown } from 'semantic-ui-react'
-import { Router } from '../../../services/routes.js'
+import { Router } from '../../services/routes.js'
 import { Mutation } from 'react-apollo';
-import { MUTATION_FRIEND_PROFILE } from '../../../services/requests/profile.js';
-import '../profile.less'
+import { MUTATION_FRIEND_PROFILE } from '../../services/requests/profile.js';
+import '../Profile/profile.less'
 
 //-- React Implementation ------------------------
